services/market-stream-service/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.43
sqlalchemy>=2.0.0
schedule>=1.2.0
kafka-python>=2.0.2
redis>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

